Refactor init_db.js to use the user schema such that property defaults are used. Create a script to perform this initialization of the developemnt database.
